[setup] Renamed directory to avoid case-sensitive problems
